requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==3.17.4
mammoth==1.6.0

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==5.1.1
transformers==4.56.2
huggingface-hub==0.35.1

# Environment and utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS and middleware
fastapi-cors==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For better performance
orjson==3.9.10
